requirements.txt
vkbottle==4.3.9
aiogram==3.0.0b4

python-dotenv==0.20.0
loguru==0.6.0
transitions==0.9.0